Fix size_t printf conversion specifiers: use %llu and cast argument to (unsigned long long)

+++ asyn/drvAsynFTDI/ftdiDriver.cpp
@@ -606,11 +606,11 @@ FTDIDriverStatus FTDIDriver::write(const unsigned char *buffer, int bufferSize, 
   if (rc[3] > 0){
     // Since we write also the address byte byteswritten = sz+1
     *bytesWritten = (size_t)(sz+1);
-    debugPrint("%s : %lu bytes written err=%d: ", functionName, *bytesWritten, err);
+    debugPrint("%s : %llu bytes written err=%d: ", functionName, (unsigned long long) *bytesWritten, err);
     if( err == FTDIDriverError ) debugPrint("FTDIDriverError\n");
     else                         debugPrint("FTDIDriverSUCCESS\n");
   } else {
-    debugPrint("FTDI write error: %lu (%s)\n", rc, ftdi_get_error_string(ftdi_));
+    debugPrint("FTDI write error: %llu (%s)\n", (unsigned long long) rc, ftdi_get_error_string(ftdi_));
     *bytesWritten = 0;
     return FTDIDriverError;
   }
@@ -640,7 +640,8 @@ FTDIDriverStatus FTDIDriver::read(unsigned char *buffer, size_t bufferSize, size
     return FTDIDriverError;
   }
 
-  debugPrint("%s : Reading: bufferSize=%lu, timeout=%d\n", functionName, bufferSize, timeout);
+  debugPrint("%s : Reading: bufferSize=%llu, timeout=%d\n",
+             functionName, (unsigned long long) bufferSize, timeout);
 
   timeval stime;
   timeval ctime;
